refactor: move restaurant data extraction logic to a dedicated function and enhance backfill workflow
